[test-runner] Make app runable on device
try to build with fastlane

run with command line

add gemfile

build apps without codesigning

seperate build steps

add sign identification

Try to manually sign

Enable gcloud

Adapt to new render-test way

update xctestrun file

try to add attachment

update test result with xctest, try to extract test result from firebase

+++ render-test/runner.cpp
@@ -666,7 +666,7 @@ uint32_t getImageTileOffset(const std::set<uint32_t>& dims, uint32_t dim) {
         assert(false);
         return 0;
     }
-    return std::distance(dims.begin(), it) * kTileSizeUint;
+    return static_cast<uint32_t>(std::distance(dims.begin(), it)) * kTileSizeUint;
 }
 
 } // namespace
